test: Add unit test for creating rows.
The test is  based on the example code used in the demo.
Because the example was rather abstract, it is removed to be
replaced with a more concrete instance.

Discovered an issue with serialization of bigints when querying with
json functions. The bigints are serialized by postgres as numbers,
which causes pg-promise to deserialize them as numbers.
As a temporary fix, the js data type of bigintsin the db schema is
set to number. This will cause loss of precision in some edge cases.